README.md
this is my first commit and I do see some large companies that used this before any permissions were granted, so please ask don't just take and hope not to get caught. On that note please ask for the licenses, please use the word for word, do not Copyright and that includes pictures for some who don't know that.